Provide an environment variable to direct tests to load binaries from.
This commit adds an environment variable, `ENTROPY_BINARY_DIR`, that
allows overwriting the location from which the test suite will look for
the `entropy` and `server` binaries. I want this so that I can easily
test different binary builds. This also makes it easier to use the same
tests in CI and other environments, such as for example building locally
on a macOS machine locally but with a binary built for GNU/Linux in CI.